add source prop to BlogPostCard
I'm using the sourceInstanceName to generate the base path for the button on each blog post's card,
to allow me to use the same card component for each section.

The assumption is that, because the filesystem source name is fixed this should provide a rigid way
to determine the path.